app/proj01/requirements.txt
scrapy
scrapy-user-agents
setuptools
psycopg2-binary